+++ GUI-dev/src/app/@theme/providers/backend-server/serverconfig.ts
@@ -7,8 +7,10 @@ export class ServerConfigProvider {
   private _host = 'http://localhost:10000/';
   private _checkIfMailTaken = `${this._host}user/register/checkIfMailTaken/`;
   private _getUser = `${this._host}/api/me/`;
+  private _getStoreProcesses = `${this._host}api/store/processes`;
 
   public get checkIfMailTaken(): string {return this._checkIfMailTaken};
   public get getUser(): string {return this._getUser};
+  public get getStoreProcesses(): string {return this._getStoreProcesses};
 
 }